Windows Classic Console API support in example.cpp
